+++ src/main/java/org/jabref/gui/preferences/export/ExportTabViewModel.java
@@ -1,8 +1,7 @@
 package org.jabref.gui.preferences.export;
 
 import java.util.ArrayList;
-import java.util.Comparator;
-import java.util.List;
+import java.util.Set;
 
 import javafx.beans.property.BooleanProperty;
 import javafx.beans.property.ListProperty;
@@ -14,6 +13,7 @@ import org.jabref.gui.commonfxcontrols.SortCriterionViewModel;
 import org.jabref.gui.preferences.PreferenceTabViewModel;
 import org.jabref.model.entry.field.Field;
 import org.jabref.model.entry.field.FieldFactory;
+import org.jabref.model.entry.field.InternalField;
 import org.jabref.model.metadata.SaveOrder;
 import org.jabref.preferences.ExportPreferences;
 
@@ -44,9 +44,12 @@ public class ExportTabViewModel implements PreferenceTabViewModel {
                                                    .map(SortCriterionViewModel::new)
                                                    .toList());
 
-        List<Field> fieldNames = new ArrayList<>(FieldFactory.getCommonFields());
-        fieldNames.sort(Comparator.comparing(Field::getDisplayName));
-        sortableFieldsProperty.addAll(fieldNames);
+        Set<Field> fields = FieldFactory.getAllFieldsWithOutInternal();
+        fields.add(InternalField.INTERNAL_ALL_FIELD);
+        fields.add(InternalField.INTERNAL_ALL_TEXT_FIELDS_FIELD);
+        fields.add(InternalField.KEY_FIELD);
+        fields.add(InternalField.TYPE_HEADER);
+        sortableFieldsProperty.addAll(FieldFactory.getStandardFieldsWithCitationKey());
     }
 
     @Override